# swagger.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Swagger Integrity-First - OpenAPI 3.0",
    "description": "This is a Intergity-First Server based on the OpenAPI 3.0 specification.",
    "contact": {
      "email": "taha.muhammad.official@gmail.com"
    },
    "version": "1.0.11"
  },
  "servers": [
    {
      "url": "http://localhost:3002"
    },
    {
      "url": "https://staging.integrity1auto.com"
    },
    {
      "url": "https://api.integrity1auto.com"
    }
  ],
...
  "paths": {
...
    "/admin/appointments": {
      "post": {
        "tags": ["Admin"],
        "summary": "Admin can create appointment",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "description": "",
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "userId": {
                    "type": "string",
                    "example": "63a22fea252ee373aa91d3ad"
                  },
                  "appointmentAt": {
                    "type": "string",
                    "example": "2022-09-20T21:58:00"
                  },
                  "description": {
                    "type": "string"
                  },
                  "isEarlyBird": {
                    "type": "boolean",
                    "example": false
                  },
                  "isAfterHours": {
                    "type": "boolean",
                    "example": false
                  },
                  "vehicle": {
                    "type": "string",
                    "example": "63a22fea252ee373aa91d3ad"
                  },
                  "appointmentReason": {
                    "type": "string",
                    "example": "63a430efade356d297d459c0"
                  }
                }
              }
            }
          },
          "required": true
        },
        "description": "",
        "responses": {
          "200": {
            "description": "Successful operation"
          }
        }
      }
    },
...
    "/web-hooks/points": {
      "post": {
        "tags": ["Webhooks"],
        "summary": "Reward points to user",
        "security": [
          {
            "WebhookAPIKeyAuth": []
          }
        ],
        "requestBody": {
          "description": "",
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "kind": {
                    "type": "string",
                    "example": "LOYALTY | REFERRAL"
                  },
                  "userTekmetricId": {
                    "type": "number",
                    "example": 27800881
                  }
                }
              }
            }
          },
          "required": true
        },
        "description": "",
        "responses": {
          "200": {
            "description": "Successful operation"
          }
        }
      }
    },
...
    "/appointment": {
      "get": {
        "tags": ["Appointment"],
        "summary": "Get the list of paginated appointments of current user",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "schema": {
              "type": "number",
              "default": 1
            }
          },
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "number",
              "default": 10
            }
          }
        ],
        "description": "",
        "responses": {
          "200": {
            "description": "Successful operation"
          }
        }
      },
      "post": {
        "tags": ["Appointment"],
        "summary": "Create appointment",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "description": "",
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "appointmentAt": {
                    "type": "string",
                    "example": "2022-09-20T21:58:00"
                  },
                  "description": {
                    "type": "string"
                  },
                  "isEarlyBird": {
                    "type": "boolean",
                    "example": false
                  },
                  "isAfterHours": {
                    "type": "boolean",
                    "example": false
                  },
                  "vehicle": {
                    "type": "string",
                    "example": "63a22fea252ee373aa91d3ad"
                  },
                  "appointmentReason": {
                    "type": "string",
                    "example": "63a430efade356d297d459c0"
                  },
                  "pageName": {
                    "type": "string",
                    "example": "SERVICE_DUE | PLATNUM_MEMBER | RECENTLY_RECOMMENDED | APPOINTMENT_TAB"
                  }
                }
              }
            }
          },
          "required": true
        },
        "description": "",
        "responses": {
          "200": {
            "description": "Successful operation"
          }
        }
      }
    },
...
    "/appointment/guest": {
      "post": {
        "tags": ["Appointment"],
        "summary": "Create guest appointment",
        "security": [],
        "requestBody": {
          "description": "",
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "tekmetricShopId": {
                    "type": "number",
                    "example": 793
                  },
                  "fullName": {
                    "type": "string",
                    "example": "Taha Muhammad"
                  },
                  "email": {
                    "type": "string",
                    "example": "taha.muhammad.official@gmail.com"
                  },
                  "phoneNumber": {
                    "type": "string",
                    "example": "+923204858850"
                  },
                  "referralCode": {
                    "type": "string",
                    "example": "(Optional)"                    
                  },
                  "appointmentAt": {
                    "type": "string",
                    "example": "2022-09-20T21:58:00"
                  },
                  "description": {
                    "type": "string"
                  },
                  "isEarlyBird": {
                    "type": "boolean",
                    "example": false
                  },
                  "isAfterHours": {
                    "type": "boolean",
                    "example": false
                  },

                  "appointmentReason": {
                    "type": "string",
                    "example": "63a430efade356d297d459c0"
                  }
                }
              }
            }
          },
          "required": true
        },
        "description": "",
        "responses": {
          "200": {
            "description": "Successful operation"
          }
        }
      }
    },
    "/appointment/{appointmentId}": {
      "get": {
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "appointmentId",
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "tags": ["Appointment"],
        "summary": "Get appointment by ID",
        "description": "",
        "responses": {
          "200": {
            "description": "Successful operation"
          }
        }
      },
      "patch": {
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "appointmentId",
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "tags": ["Appointment"],
        "summary": "Update appointment",
        "description": "",
        "requestBody": {
          "description": "",
          "content": {
            "application/json": {
              "schema": {
                "type": "object"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful operation"
          }
        }
      }
    },
    "/appointment/{appointmentIds}": {
      "delete": {
        "tags": ["Appointment"],
        "summary": "Delete multiple appointments",
        "parameters": [
          {
            "name": "appointmentIds",
            "in": "path",
            "schema": {
              "type": "string",
              "example": "64065fe7f8db2e031d80fcc5,64065ffef8db2e031d80fd60"
            }
          }
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "description": "",
        "responses": {
          "200": {
            "description": "Successful operation"
          }
        }
      }
    },
...
  },
  "components": {
...
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      },
      "WebhookAPIKeyAuth": {
        "type": "apiKey",
        "name": "Authorization",
        "in": "header"
      }
    }
  }
}